feat: Cache Cohort dependencies
This change introduces caching to forward and reverse dependency lookups
for Cohorts. Identifying which Cohorts depend on any other given Cohort
is an expensive calculation because it requires checking the properties
of every other Cohort in the team.

This change introduces cache keys for fast lookups of dependencies and
dependents for any given Cohort.